requirements.txt
Flask==3.0.0
cryptography==41.0.7
requests==2.31.0
paho-mqtt==1.6.1
python-dotenv==1.0.0
sseclient-py==1.8.0
